splitting: make split private
