+++ src/lenskit/splitting/_split.py
